+++ R/h_survival_biomarkers_subgroups.R
@@ -1,0 +1,256 @@
+#' Helper Functions for Tabulating Biomarker Effects on Survival by Subgroup
+#'
+#' Helper functions which are documented here separately to not confuse the user
+#' when reading about the user-facing functions.
+#'
+#' @inheritParams survival_biomarkers_subgroups
+#' @inheritParams argument_convention
+#' @inheritParams fit_coxreg_multivar
+#' @name h_survival_biomarkers_subgroups
+#' @order 1
+#' @examples
+#' # Testing dataset.
+#' library(scda)
+#' library(dplyr)
+#' library(forcats)
+#' library(rtables)
+#'
+#' adtte <- synthetic_cdisc_data("latest")$adtte
+#'
+#' # Save variable labels before data processing steps.
+#' adtte_labels <- var_labels(adtte)
+#'
+#' adtte_f <- adtte %>%
+#'   filter(PARAMCD == "OS") %>%
+#'   mutate(
+#'     AVALU = as.character(AVALU),
+#'     is_event = CNSR == 0
+#'   ) %>%
+#'   var_relabel(
+#'     AVALU = adtte_labels["AVALU"],
+#'     is_event = "Event Flag"
+#'   )
+#'
+NULL
+
+#' @describeIn h_survival_biomarkers_subgroups helps with converting the "survival" function variable list
+#'   to the "Cox regression" variable list. The reason is that currently there is an inconsistency between the variable
+#'   names accepted by `extract_survival_subgroups()` and `fit_coxreg_multivar()`.
+#' @param biomarker (`string`)\cr the name of the biomarker variable.
+#' @export
+#' @examples
+#' # This is how the variable list is converted internally.
+#' h_surv_to_coxreg_variables(
+#'   variables = list(
+#'     tte = "AVAL",
+#'     is_event = "EVNT",
+#'     covariates = c("A", "B"),
+#'     strata = "D"
+#'   ),
+#'   biomarker = "AGE"
+#' )
+h_surv_to_coxreg_variables <- function(variables, biomarker) {
+  assert_that(
+    is.list(variables),
+    is.string(variables$tte),
+    is.string(variables$is_event),
+    is.string(biomarker)
+  )
+  list(
+    time = variables$tte,
+    event = variables$is_event,
+    arm = biomarker,
+    covariates = variables$covariates,
+    strata = variables$strata
+  )
+}
+
+#' @describeIn h_survival_biomarkers_subgroups prepares estimates for number of events, patients and median survival
+#'   times, as well as hazard ratio estimates, confidence intervals and p-values, for multiple biomarkers
+#'   in a given single data set.
+#'   `variables` corresponds to names of variables found in `data`, passed as a named list and requires elements
+#'   `tte`, `is_event`, `biomarkers` (vector of continuous biomarker variables) and optionally `subgroups` and `strat`.
+#' @export
+#' @examples
+#'
+#' # For a single population, estimate separately the effects
+#' # of two biomarkers.
+#' df <- h_coxreg_mult_cont_df(
+#'   variables = list(
+#'     tte = "AVAL",
+#'     is_event = "is_event",
+#'     biomarkers = c("BMRKR1", "AGE"),
+#'     covariates = "SEX",
+#'     strata = c("STRATA1", "STRATA2")
+#'   ),
+#'   data = adtte_f
+#' )
+#' df
+#'
+#' # If the data set is empty, still the corresponding rows with missings are returned.
+#' h_coxreg_mult_cont_df(
+#'   variables = list(
+#'     tte = "AVAL",
+#'     is_event = "is_event",
+#'     biomarkers = c("BMRKR1", "AGE"),
+#'     covariates = "REGION1",
+#'     strata = c("STRATA1", "STRATA2")
+#'   ),
+#'   data = adtte_f[NULL, ]
+#' )
+h_coxreg_mult_cont_df <- function(variables,
+                                  data,
+                                  control = control_coxreg()) {
+  assert_that(
+    is.list(variables),
+    is_character_vector(variables$biomarkers),
+    is_df_with_variables(data, as.list(unlist(variables))),
+    is_fully_named_list(control)
+  )
+  conf_level <- control[["conf_level"]]
+  pval_label <- paste0(
+    "p-value (", stringr::str_to_sentence(control[["pval_method"]]), ")"
+  )
+  # If there is any data, run model, otherwise return empty results.
+  if (nrow(data) > 0) {
+    bm_cols <- match(variables$biomarkers, names(data))
+    l_result <- lapply(variables$biomarkers, function(bm) {
+      coxreg_list <- fit_coxreg_multivar(
+        variables = h_surv_to_coxreg_variables(variables, bm),
+        data = data,
+        control = control
+      )
+      result <- do.call(
+        h_coxreg_multivar_extract,
+        c(list(var = bm), coxreg_list[c("mod", "data", "control")])
+      )
+      data_fit <- as.data.frame(as.matrix(coxreg_list$mod$y))
+      data_fit$status <- as.logical(data_fit$status)
+      median <- s_surv_time(
+        df = data_fit,
+        .var = "time",
+        is_event = "status"
+      )$median
+      data.frame(
+        # Dummy column needed downstream to create a nested header.
+        biomarker = bm,
+        biomarker_label = var_labels(data[bm], fill = TRUE),
+        n_tot = coxreg_list$mod$n,
+        n_tot_events = coxreg_list$mod$nevent,
+        median = as.numeric(median),
+        result[1L, c("hr", "lcl", "ucl")],
+        conf_level = conf_level,
+        pval = result[1L, "pval"],
+        pval_label = pval_label,
+        stringsAsFactors = FALSE
+      )
+    })
+    do.call(rbind, args = c(l_result, make.row.names = FALSE))
+  } else {
+    data.frame(
+      biomarker = variables$biomarkers,
+      biomarker_label = var_labels(data[variables$biomarkers], fill = TRUE),
+      n_tot = 0L,
+      n_tot_events = 0L,
+      median = NA,
+      hr = NA,
+      lcl = NA,
+      ucl = NA,
+      conf_level = conf_level,
+      pval = NA,
+      pval_label = pval_label,
+      row.names = seq_along(variables$biomarkers),
+      stringsAsFactors = FALSE
+    )
+  }
+}
+
+#' @describeIn h_survival_biomarkers_subgroups prepares a single sub-table given a `df_sub` containing
+#'   the results for a single biomarker.
+#' @param df (`data.frame`)\cr results for a single biomarker, as part of what is
+#'   returned by [extract_survival_biomarkers()] (it needs a couple of columns which are
+#'   added by that high-level function relative to what is returned by [h_coxreg_mult_cont_df()],
+#'   see the example).
+#' @export
+#' @examples
+#'
+#' # Starting from above `df`, zoom in on one biomarker and add required columns.
+#' df1 <- df[1, ]
+#' df1$subgroup <- "All patients"
+#' df1$row_type <- "content"
+#' df1$var <- "ALL"
+#' df1$var_label <- "All patients"
+#' h_tab_surv_one_biomarker(
+#'   df1,
+#'   vars = c("n_tot", "n_tot_events", "median", "hr", "ci", "pval"),
+#'   time_unit = "days"
+#' )
+h_tab_surv_one_biomarker <- function(df,
+                                     vars,
+                                     time_unit) {
+  conf_level <- df$conf_level[1]
+  method <-  df$pval_label[1]
+  afuns <- a_survival_subgroups()[vars]
+  colvars <- d_survival_subgroups_colvars(
+    vars,
+    conf_level = conf_level,
+    method = method,
+    time_unit = time_unit
+  )
+
+  lyt <- basic_table()
+
+  # Row split by row type - only keep the content rows here.
+  lyt <- split_rows_by(
+    lyt = lyt,
+    var = "row_type",
+    split_fun = keep_split_levels("content"),
+    nested = FALSE,
+    indent_mod = 1L
+  )
+
+  # Summarize rows with all patients.
+  lyt <- summarize_row_groups(
+    lyt = lyt,
+    var = "var_label",
+    cfun = afuns
+  )
+
+  # Split cols by the multiple variables to populate into columns.
+  lyt <- split_cols_by_multivar(
+    lyt = lyt,
+    vars = colvars$vars,
+    varlabels = colvars$labels
+  )
+
+  # If there is any subgroup variables, we extend the layout accordingly.
+  if ("analysis" %in% df$row_type) {
+
+    # Now only continue with the subgroup rows.
+    lyt <- split_rows_by(
+      lyt = lyt,
+      var = "row_type",
+      split_fun = keep_split_levels("analysis"),
+      nested = FALSE,
+      child_labels = "hidden"
+    )
+
+    # Split by the subgroup variable.
+    lyt <- split_rows_by(
+      lyt = lyt,
+      var = "var",
+      labels_var = "var_label",
+      nested = TRUE,
+      child_labels = "visible",
+      indent_mod = 1L
+    )
+
+    # Then analyze colvars for each subgroup.
+    lyt <- summarize_row_groups(
+      lyt = lyt,
+      cfun = afuns,
+      var = "subgroup"
+    )
+  }
+  build_table(lyt, df = df)
+}